requirements.txt
pandas
flask
gunicorn
python-binance
